Rename variable 'name' to 'dnpName' when refering to dnpName

+++ packages/dappmanager/src/modules/legacy/migrateLegacyEnvFiles.ts
@@ -15,29 +15,43 @@ export async function migrateLegacyEnvFiles(
   dnpList: PackageContainer[]
 ): Promise<void> {
   try {
-    for (const { name, isCore } of dnpList) migrateLegacyEnvFile(name, isCore);
+    for (const { dnpName, isCore } of dnpList)
+      migrateLegacyEnvFile(dnpName, isCore);
     logs.info("Finished migrating legacy DNP .env files if any");
   } catch (e) {
     logs.error("Error migrating DNP .env files", e);
   }
 }
 
-export function migrateLegacyEnvFile(name: string, isCore: boolean): boolean {
-  const envFilePath = getPath.envFile(name, isCore);
+export function migrateLegacyEnvFile(
+  dnpName: string,
+  isCore: boolean
+): boolean {
+  const envFilePath = getPath.envFile(dnpName, isCore);
   try {
     const envFileData = fs.readFileSync(envFilePath, "utf8");
     const envsArray = envFileData.trim().split("\n");
 
-    const compose = new ComposeFileEditor(name, isCore);
-    compose.service().mergeEnvs(parseEnvironment(envsArray));
-    compose.service().omitDnpEnvFile();
-    compose.write();
+    const compose = new ComposeFileEditor(dnpName, isCore);
+    const singleService = compose
+      .services()
+      .find(service => service.serviceName === dnpName);
+    if (singleService) {
+      singleService.mergeEnvs(parseEnvironment(envsArray));
+      singleService.omitDnpEnvFile();
+      compose.write();
 
-    fs.unlinkSync(envFilePath);
-    logs.info(`Converted ${name} .env file to compose environment`);
-    return true;
+      fs.unlinkSync(envFilePath);
+      logs.info(`Converted ${dnpName} .env file to compose environment`);
+      return true;
+    } else {
+      throw Error(
+        `Can not migrate ENVs for multi-service packages: ${dnpName}`
+      );
+    }
   } catch (e) {
-    if (!isNotFoundError(e)) logs.error(`Error migrating ${name} .env file`, e);
+    if (!isNotFoundError(e))
+      logs.error(`Error migrating ${dnpName} .env file`, e);
     return false;
   }
 }
